requirements.txt
playwright==1.20.1
pytest==7.1.1
requests
allure-pytest
ipdb
python-dotenv
